+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 ## Unreleased
 
 ### Changed
-- Changed `\dukt\social\github\loginproviders\Github::getOauthProvider()` method’s visibility to `public`.
+- Changed `dukt\social\github\loginproviders\Github::getOauthProvider()` method’s visibility to `public`.
+- Removed `dukt\social\github\loginproviders\Github::getProfile()`.
 
 ## 3.0.2 - 2018-05-18